Modification on Deposit and Withdraw functions
This file version adds a list to the class's data attributes and modifies the deposit and withdraw method to add transaction details to the list. This file will implement the 'pytz' and 'datetime' module for local time to capture the date and time of the transaction as well as the amount.